Move facilities
